requirements.txt
discord.py
discord-components
python-dotenv